telegram_bot_final_with_token/requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.120.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0